qcreport: rename 'screens.py' module to 'fastq_screen.py'.
